rename, restructure, styling

+++ src/core/constants.rs
@@ -1,3 +1,3 @@
 pub const GET_RESERVES_SELECTOR: &str =
     "0x3cb0e1486e633fbe3e2fafe8aedf12b70ca1860e7467ddb75a17858cde39312";
-pub const SCALE: f64 = 1000000 as f64;
+pub const SCALE: f64 = 1000000_f64;
